review godocs; drop superfluous 'conn' prefix from files; rename methods.
