+++ connmgr/manager.go
