+++ control/disconnect.go
@@ -1,4 +1,9 @@
 package control
 
-// DisconnectReason informs the receiver of the reason the connectedness is being terminated by the sender.
+// DisconnectReason communicates the reason why a connection is being closed.
+//
+// A zero value stands for "no reason" / NA.
+//
+// This is an EXPERIMENTAL type. It will change in the future. Refer to the
+// connmgr.ConnectionGater godoc for more info.
 type DisconnectReason int
